WIP feature MFA TOTP/HOTP

See merge request tine20/tine20!611

+++ tine20/Tinebase/Model/MFA/HOTPUserConfig.php
@@ -9,26 +9,25 @@
  * @author      Paul Mehrer <p.mehrer@metaways.de>
  */
 
+use ParagonIE\ConstantTime\Base32;
+
+
 /**
- * Yubico OTP MFA UserConfig Model
+ * HOTP MFA UserConfig Model
  *
  * @package     Tinebase
  * @subpackage  Auth
  */
-class Tinebase_Model_MFA_YubicoOTPUserConfig extends Tinebase_Auth_MFA_AbstractUserConfig
+class Tinebase_Model_MFA_HOTPUserConfig extends Tinebase_Auth_MFA_AbstractUserConfig
 {
-    public const MODEL_NAME_PART = 'MFA_YubicoOTPUserConfig';
+    public const MODEL_NAME_PART = 'MFA_HOTPUserConfig';
 
-    public const FLD_AES_KEY = 'aes_key';
-    public const FLD_CC_ID = 'cc_id';
-    public const FLD_PRIVAT_ID = 'private_id';
-    public const FLD_PUBLIC_ID = 'public_id';
     public const FLD_ACCOUNT_ID = 'account_id';
+    public const FLD_CC_ID = 'cc_id';
     public const FLD_COUNTER = 'counter';
-    public const FLD_SESSIONC = 'sessionc';
+    public const FLD_SECRET = 'secret';
 
-    protected $_aesKey;
-    protected $_privatId;
+    protected $_secret;
 
     /**
      * Holds the model configuration (must be assigned in the concrete class)
@@ -38,39 +37,28 @@ class Tinebase_Model_MFA_YubicoOTPUserConfig extends Tinebase_Auth_MFA_AbstractU
     protected static $_modelConfiguration = [
         self::APP_NAME                      => Tinebase_Config::APP_NAME,
         self::MODEL_NAME                    => self::MODEL_NAME_PART,
-        self::RECORD_NAME                   => 'Yubico OTP',
-        self::TITLE_PROPERTY                => 'Yubico OTP',
+        self::RECORD_NAME                   => 'Counter based OTP (HOTP)',
+        self::RECORDS_NAME                  => 'Counter based OTPs (HOTP)', // ngettext('Counter based OTP (HOTP)', 'Counter based OTPs (HOTP)', n)
+        self::TITLE_PROPERTY                => 'Counter based OTP (HOPT) is configured', // _('Counter based OTP (HOTP) is configured')
 
         self::FIELDS                        => [
-            self::FLD_PUBLIC_ID                 => [
+            self::FLD_ACCOUNT_ID                => [
                 self::TYPE                          => self::TYPE_STRING,
-                self::LABEL                         => 'Yubico OTP public id', // _('Yubico OTP public id')
-                self::VALIDATORS                    => [
-                    Zend_Filter_Input::ALLOW_EMPTY => false,
-                    Zend_Filter_Input::PRESENCE => Zend_Filter_Input::PRESENCE_REQUIRED,
-                ],
+                self::DISABLED                      => true,
             ],
-            self::FLD_PRIVAT_ID                 => [
+            self::ID                            => [
                 self::TYPE                          => self::TYPE_STRING,
-                self::LABEL                         => 'Yubico OTP privat id', // _('Yubico OTP privat id')
+                self::DISABLED                      => true,
             ],
-            self::FLD_AES_KEY                   => [
+            self::FLD_SECRET                    => [
                 self::TYPE                          => self::TYPE_STRING,
-                self::LABEL                         => 'Yubico privat key', // _('Yubico privat key')
+                self::LABEL                         => 'Secret Key', // _('Secret Key')
             ],
             self::FLD_CC_ID                     => [
                 self::TYPE                          => self::TYPE_STRING,
                 self::DISABLED                      => true,
             ],
-            self::FLD_ACCOUNT_ID                => [
-                self::TYPE                          => self::TYPE_STRING,
-                self::DISABLED                      => true,
-            ],
             self::FLD_COUNTER                   => [
-                self::TYPE                          => self::TYPE_INTEGER,
-                self::DISABLED                      => true,
-            ],
-            self::FLD_SESSIONC                  => [
                 self::TYPE                          => self::TYPE_INTEGER,
                 self::DISABLED                      => true,
             ],
@@ -86,44 +74,41 @@ class Tinebase_Model_MFA_YubicoOTPUserConfig extends Tinebase_Auth_MFA_AbstractU
 
     public function setFromArray(array &$_data)
     {
-        if (isset($_data[self::FLD_AES_KEY])) {
-            $this->_aesKey = $_data[self::FLD_AES_KEY];
-            unset($_data[self::FLD_AES_KEY]);
-        }
-        if (isset($_data[self::FLD_PRIVAT_ID])) {
-            $this->_privatId = $_data[self::FLD_PRIVAT_ID];
-            unset($_data[self::FLD_PRIVAT_ID]);
+        if (isset($_data[self::FLD_SECRET])) {
+            $this->_secret = $_data[self::FLD_SECRET];
+            unset($_data[self::FLD_SECRET]);
         }
         parent::setFromArray($_data);
     }
 
-    public function getAesKey(): ?string
+    public function getSecret(): ?string
     {
-        return $this->_aesKey;
-    }
-
-    public function getPrivatId(): ?string
-    {
-        return $this->_privatId;
+        return $this->_secret;
     }
 
     public function updateUserNewRecordCallback(Tinebase_Model_FullUser $newUser, Tinebase_Model_FullUser $oldUser, Tinebase_Model_MFA_UserConfig $userCfg)
     {
         $this->{self::FLD_ACCOUNT_ID} = $newUser->getId();
-        if (($newAes = $this->getAesKey()) && $newId = $this->getPrivatId()) {
+        if (($newSecret = $this->getSecret())) {
+            if (preg_match('/[^A-Z2-7]/', $newSecret)) {
+                throw new Tinebase_Exception_UnexpectedValue('secret needs to be base32 conform, consisting only of A-Z + 2-7 chars');
+            }
+            if (!$this->{self::ID}) {
+                $this->{self::ID} = Tinebase_Record_Abstract::generateUID();
+            }
             $cc = Tinebase_Auth_CredentialCache::getInstance();
             $adapter = explode('_', get_class($cc->getCacheAdapter()));
             $adapter = end($adapter);
             try {
                 $cc->setCacheAdapter('Shared');
-                $sharedCredentials = Tinebase_Auth_CredentialCache::getInstance()->cacheCredentials($newId,
-                    $newAes, null, true /* save in DB */, Tinebase_DateTime::now()->addYear(100));
+                $sharedCredentials = Tinebase_Auth_CredentialCache::getInstance()->cacheCredentials($this->{self::ID},
+                    $newSecret, null, true /* save in DB */, Tinebase_DateTime::now()->addYear(100));
 
                 $this->{self::FLD_CC_ID} = $sharedCredentials->getId();
 
                 if ($oldUser->mfa_configs && ($oldCfg = $oldUser->mfa_configs->find(Tinebase_Model_MFA_UserConfig::FLD_ID,
                         $userCfg->{Tinebase_Model_MFA_UserConfig::FLD_ID})) && ($ccId = $oldCfg
-                        ->{Tinebase_Model_MFA_UserConfig::FLD_CONFIG}->{self::FLD_CC_ID})) {
+                        ->{Tinebase_Model_MFA_UserConfig::FLD_CONFIG}->{self::FLD_CC_ID}) && $ccId !== $this->{self::FLD_CC_ID}) {
                     $cc->delete($ccId);
                 }
             } finally {
